refactor: dom_types
