+++ src/dom_entity_names.rs
@@ -1,0 +1,3 @@
+pub mod attributes;
+pub mod styles;
+pub mod tags;
