--- src/dom_entity_names/mod.rs
@@ -1,1 +1,0 @@
-pub mod styles;
