+++ src/dom_entity_names/tags/tag_names.rs
@@ -1,0 +1,122 @@
+// Comprehensive lists:
+// - https://developer.mozilla.org/en-US/docs/Web/HTML/Element
+// - https://developer.mozilla.org/en-US/docs/Web/SVG/Element
+// Grouped here by category on Mozilla's pages, linked above.
+make_tags! {
+    // -------- Standard HTML Tags -------- //
+
+    Address => "address", Article => "article", Aside => "aside", Footer => "footer",
+    Header => "header", H1 => "h1",
+    H2 => "h2", H3 => "h3", H4 => "h4", H5 => "h5", H6 => "h6",
+    Hgroup => "hgroup", Main => "main", Nav => "nav", Section => "section",
+
+    BlockQuote => "blockquote",
+    Dd => "dd", Dir => "dir", Div => "div", Dl => "dl", Dt => "dt", FigCaption => "figcaption", Figure => "figure",
+    Hr => "hr", Li => "li", Ol => "ol", P => "p", Pre => "pre", Ul => "ul",
+
+    A => "a", Abbr => "abbr",
+    B => "b", Bdi => "bdi", Bdo => "bdo", Br => "br", Cite => "cite", Code => "code", Data => "data",
+    Dfn => "dfn", Em => "em", I => "i", Kbd => "kbd", Mark => "mark", Q => "q", Rb => "rb",
+    Rp => "rp", Rt => "rt", Rtc => "rtc", Ruby => "ruby", S => "s", Samp => "samp", Small => "small",
+    Span => "span", Strong => "strong", Sub => "sub", Sup => "sup", Time => "time", Tt => "tt",
+    U => "u", Var => "var", Wbr => "wbr",
+
+    Area => "area", Audio => "audio", Img => "img", Map => "map", Track => "track", Video => "video",
+
+    Applet => "applet", Embed => "embed", Iframe => "iframe",
+    NoEmbed => "noembed", Object => "object", Param => "param", Picture => "picture", Source => "source",
+
+    Canvas => "canvas", NoScript => "noscript", Script => "Script",
+
+    Del => "del", Ins => "ins",
+
+    Caption => "caption", Col => "col", ColGroup => "colgroup", Table => "table", Tbody => "tbody",
+    Td => "td", Tfoot =>"tfoot", Th => "th", Thead => "thead", Tr => "tr",
+
+    Button => "button", DataList => "datalist", FieldSet => "fieldset", Form => "form", Input => "input",
+    Label => "label", Legend => "legend", Meter => "meter", OptGroup => "optgroup", Option => "option",
+    Output => "output", Progress => "progress", Select => "select", TextArea => "textarea",
+
+    Details => "details", Dialog => "dialog", Menu => "menu", MenuItem => "menuitem", Summary => "summary",
+
+    Content => "content", Element => "element", Shadow => "shadow", Slot => "slot", Template => "template",
+
+    // -------- SVG Tags -------- //
+
+    // Animation elements
+    Animate => "animate", AnimateColor => "animateColor", AnimateMotion => "animateMotion",
+    AnimateTransform => "animateTransform", Discard => "discard", Mpath => "mpath", Set => "set",
+
+    // Shape elements
+    Circle => "circle", Ellipse => "ellipse", Line => "line", Polygon => "polygon",
+    Polyline => "polyline", Rect => "rect", Mesh => "mesh", Path => "path",
+
+    // Container elements
+    Defs => "defs", G => "g", Marker => "marker", Mask => "mask", MissingGlyph => "missing-glyph",
+    Pattern => "pattern", Svg => "svg", Switch => "switch", Symbol => "symbol", Unknown => "unknown",
+
+    // Descriptive elements
+    Desc => "desc", Metadata => "metadata", Title => "title",
+
+    // Filter primitive elements
+    FeBlend             => "feBlend",
+    FeColorMatrix       => "feColorMatrix",
+    FeComponentTransfer => "feComponentTransfer",
+    FeComposite         => "feComposite",
+    FeConvolveMatrix    => "feConvolveMatrix",
+    FeDiffuseLighting   => "feDiffuseLighting",
+    FeDisplacementMap   => "feDisplacementMap",
+    FeDropShadow        => "feDropShadow",
+    FeFlood             => "feFlood",
+    FeFuncA             => "feFuncA",
+    FeFuncB             => "feFuncB",
+    FeFuncG             => "feFuncG",
+    FeFuncR             => "feFuncR",
+    FeGaussianBlur      => "feGaussianBlur",
+    FeImage             => "feImage",
+    FeMerge             => "feMerge",
+    FeMergeNode         => "feMergeNode",
+    FeMorphology        => "feMorphology",
+    FeOffset            => "feOffset",
+    FeSpecularLighting  => "feSpecularLighting",
+    FeTile              => "feTile",
+    FeTurbulence        => "feTurbulence",
+
+    // Light source elements
+    FeDistantLight => "feDistantLight", FePointLight => "fePointLight",  FeSpotLight => "feSpotLight",
+
+    // Font elements
+    Font => "font",
+    FontFace => "font-face",
+    FontFaceFormat => "font-face-format",
+    FontFaceName => "font-face-name",
+    FontFaceSrc => "font-face-src",
+    FontFaceUri => "font-face-uri",
+    HKern => "hkern",
+    VKern => "vkern",
+
+    // Gradient elements
+    LinearGradient => "linearGradient", MeshGradient => "meshGradient",
+    RadialGradient => "radialGradient", Stop => "stop",
+
+    // Graphics elements
+    Image => "image",
+
+    // Graphics referencing elements
+    Use => "use",
+
+    // Paint server elements
+    Hatch => "hatch", SolidColor => "solidcolor",
+
+    // Text content elements
+    AltGlyph => "altGlyph", AltGlyphDef => "altGlyphDef", AltGlyphItem => "altGlyphItem", Glyph => "glyph",
+    GlyphRef => "glyphRef", TextPath => "textPath", Text => "text", TRef => "tref", TSpan => "tspan",
+
+    // Uncategorized elements
+    ClipPath => "clipPath", ColorProfile => "color-profile", Cursor => "cursor", Filter => "filter",
+    ForeignObject => "foreignObject", HatchPath => "hatchpath", MeshPatch => "meshpatch", MeshRow => "meshrow",
+    Style => "style", View => "view",
+
+    // A custom placeholder tag, for internal use
+    Placeholder => "placeholder"
+}
