+++ src/vdom/builder/init.rs
@@ -1,5 +1,5 @@
 use crate::{
-    dom_types::View,
+    dom_types::view::View,
     orders::OrdersContainer,
     routing::Url,
     vdom::builder::{
